Remove use of RegEx for splitting folders
Instead of blindly chunking a path using the directory symbol, we
want to split it into it's actual components and we have an existing
function to do that.

Also fixed is the recent menu crashing if it's not found in the menu.ui

And the prepending on .. to every recent file, no matter if it was in
the root directory or not. The code used to assume size 2 meant 1 folder
and 1 filename, but it's actually 1 root and 1 filename, so change the
checks to take advantage of that.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/5169

+++ src/ui/desktop/menubar.cpp
@@ -44,6 +44,7 @@
 #include "inkscape-application.h" // Open recent
 #include "preferences.h"          // Use icons or not
 #include "io/resource.h"          // UI File location
+#include "io/fix-broken-links.h"
 
 // =================== Main Menu ================
 void
@@ -132,9 +133,15 @@ build_menu()
         auto recent_object = refBuilder->get_object("recent-files");
         auto recent_gmenu = Glib::RefPtr<Gio::Menu>::cast_dynamic(recent_object);
         auto recent_menu_quark = Glib::Quark("recent-manager");
-        recent_gmenu->set_data(recent_menu_quark, recent_manager.get()); // mark submenu, so we can find it
+        if (recent_gmenu) {
+            recent_gmenu->set_data(recent_menu_quark, recent_manager.get()); // mark submenu, so we can find it
+        }
 
         auto rebuild_recent = [](Glib::RefPtr<Gio::Menu> const &submenu) {
+            if (!submenu) {
+                g_warning("No recent submenu in menus.ui found.");
+                return;
+            }
 
             submenu->remove_all();
 
@@ -193,9 +200,9 @@ build_menu()
                     display_uris.emplace_back(( * it   )->get_uri_display());
                     display_uris.emplace_back(( *(it+1))->get_uri_display());
 
-                    std::vector<std::vector<Glib::ustring>> path_parts;
-                    path_parts.emplace_back(Glib::Regex::split_simple(G_DIR_SEPARATOR_S, (* it   )->get_uri_display()));
-                    path_parts.emplace_back(Glib::Regex::split_simple(G_DIR_SEPARATOR_S, (*(it+1))->get_uri_display()));
+                    std::vector<std::vector<std::string>> path_parts;
+                    path_parts.emplace_back(Inkscape::splitPath((* it   )->get_uri_display()));
+                    path_parts.emplace_back(Inkscape::splitPath((*(it+1))->get_uri_display()));
 
                     // Find first directory difference from root down.
                     auto max_size = std::min(path_parts[0].size(), path_parts[1].size());
@@ -210,18 +217,15 @@ build_menu()
                     for (auto j = 0; j < 2; j++) {
 
                         auto display_uri = display_uris[j]; // We always use display_uri as map index.
+                        // Size is always one first element such as '/' or 'C:\\' and the last element is the filename
                         auto size = path_parts[j].size();
 
-                        if (size == 1 ||
-                            size == 2) {
+                        if (size <= 3) {
                             // If file is in root directory or child of root directory, just use display uri.
                             shortened_path_map[display_uri] = display_uri;
                         } else if (i == (size-1)) {
                             // If difference is at last path part (file name), use that.
-                            shortened_path_map[display_uri] =
-                                Glib::ustring::compose("..%1%2",
-                                                       G_DIR_SEPARATOR_S,
-                                                       path_parts[j].back());
+                            shortened_path_map[display_uri] = path_parts[j].back();
                         } else if (i == (size-2)) {
 
                             // If difference is last directory level (file name), use that + file name.
@@ -229,6 +233,15 @@ build_menu()
                                 Glib::ustring::compose ("..%1%2%3%4",
                                                         G_DIR_SEPARATOR_S,
                                                         path_parts[j][size-2],
+                                                        G_DIR_SEPARATOR_S,
+                                                        path_parts[j][size-1]);
+                        } else if (i == 1) {
+                            // parts[j][i] is actually a root folder
+                            shortened_path_map[display_uri] =
+                                Glib::ustring::compose ("%1%2%3..%4%5",
+                                                        path_parts[j][0],
+                                                        path_parts[j][i],
+                                                        G_DIR_SEPARATOR_S,
                                                         G_DIR_SEPARATOR_S,
                                                         path_parts[j][size-1]);
                         } else {
